requirements.txt
# clinical_pipeline_v12 — Python dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# MongoDB
pymongo==4.7.2

# Config
pyyaml==6.0.1

# HTTP client (for Ollama API calls — Step 2)
httpx==0.27.0

# Fuzzy + phonetic name matching (Step 4)
rapidfuzz==3.9.3
jellyfish==1.0.3      # Soundex / Metaphone phonetic algorithms

# Data validation
pydantic==2.7.1

# Utilities
python-dotenv==1.0.1  # Optional: load .env for secrets
